requirements.txt
Django==5.0.1
matplotlib==3.8.2
numpy==1.26.3
pandas==2.2.0
Pillow==10.2.0
scikit_learn==1.2.1
seaborn==0.13.1
djangorestframework
